feat(employer-cost-summary): remove unnecesary methods in payrollrepository, add a new query to get the gross salary for payroll. add new query builder in program.cs
