+++ backend/Application/IGetGrossSalaryByPayrollIdQuery.cs
@@ -1,0 +1,4 @@
+public interface IGetGrossSalaryByPayrollIdQuery
+{
+    decimal Execute(Guid payrollId);
+}
