+++ backend/Infraestructure/IPayrollRepository.cs
@@ -8,7 +8,6 @@ namespace backend.Infraestructure
         Task<List<PayrollModel>> GetByCompanyIdAsync(Guid companyId);
         Task<List<PayrollSummaryDto>> GetSummaryByCompanyIdAsync(Guid companyId);
         Task<bool> CheckPayrollExistsAsync(Guid companyId, DateTime startDate, DateTime endDate);
-        decimal GetTotalGrossSalaryByPayrollId(Guid payrollId);
 
     }
 }
