+++ backend/Infraestructure/PayrollRepository.cs
@@ -138,26 +138,5 @@ namespace backend.Infraestructure
                 _connection.Close();
             }
         }
-        public decimal GetTotalGrossSalaryByPayrollId(Guid payrollId)
-        {
-            var total = 0m;
-
-            using var connection = new SqlConnection(_connectionString);
-            var query = @"
-                SELECT SUM(GrossSalary)
-                FROM PaymentDetails
-                WHERE PayrollId = @PayrollId";
-
-            using var command = new SqlCommand(query, connection);
-            command.Parameters.AddWithValue("@PayrollId", payrollId);
-
-            connection.Open();
-            var result = command.ExecuteScalar();
-            if (result != DBNull.Value)
-                total = Convert.ToDecimal(result);
-
-            return total;
-        }
-
     }
 }
